Plutarch.LedgerApi.Interval

Description

This module is meant to be imported qualified, as some of its identifiers clash with the Plutarch prelude, as well as other parts of the Plutarch API.

Types

newtype PInterval (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PInterval (Term s (PDataRecord '["from" ':= PLowerBound a, "to" ':= PUpperBound a]))

newtype PLowerBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PLowerBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

newtype PUpperBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PUpperBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

data PExtended (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" ':= a]))

PPosInf (Term s (PDataRecord '[]))

Functions

Creation

psingleton :: forall (a :: S -> Type) (s :: S). Term s (PAsData a :--> PInterval a)

Given x, create the interval [x, x].

Since: 2.1.1

pfrom :: forall (a :: S -> Type) (s :: S). Term s (PAsData a :--> PInterval a)

Given x, create the interval [x, +infty)

Since: 2.1.1

pto :: forall (a :: S -> Type) (s :: S). Term s (PAsData a :--> PInterval a)

Given x, create the interval (-infty, x].

Since: 2.1.1

palways :: forall (a :: S -> Type) (s :: S). (FromData (AsHaskell a), ToData (AsHaskell a)) => Term s (PInterval a)

Create the interval (-infty, +infty).

Since: 2.1.1

pinterval :: forall (a :: S -> Type) (s :: S). Term s (PAsData a :--> (PAsData a :--> PInterval a))

pinterval x y creates the interval [x, y].

Since: 2.1.1

pinclusiveLowerBound :: forall (a :: S -> Type) (s :: S). (PIsData a, PCountable a) => Term s (PLowerBound a :--> PExtended a)

Turn a PLowerBound into a single inclusive bounding value.

@since WIP

pinclusiveUpperBound :: forall (a :: S -> Type) (s :: S). (PIsData a, PEnumerable a) => Term s (PUpperBound a :--> PExtended a)

Turn a PUpperBound into a single inclusive bounding value.

@since WIP

Queries

pmember :: forall (a :: S -> Type) (s :: S). (PIsData a, PEnumerable a) => Term s (PAsData a :--> (PInterval a :--> PBool))

Check if a value is inside the given interval.

@since WIP

pcontains :: forall (a :: S -> Type) (s :: S). (PIsData a, PEnumerable a) => Term s (PInterval a :--> (PInterval a :--> PBool))

pcontains # i1 # i2 is true if i2 is entirely contained in i1: more specifically, if for any s, if pmember # s # i2, then pmember # s # i1.

@since WIP

pbefore :: forall (a :: S -> Type) (s :: S). (POrd a, PIsData a) => Term s (a :--> (PInterval a :--> PBool))

before x i is true if x is earlier than the start of i.

Since: 2.1.1

pafter :: forall (a :: S -> Type) (s :: S). (POrd a, PIsData a) => Term s (a :--> (PInterval a :--> PBool))

after x u is true if x is later than the end of i.

Since: 2.1.1

pisEmpty :: forall (a :: S -> Type) (s :: S). (PIsData a, PEnumerable a) => Term s (PInterval a :--> PBool)

Check if a PInterval is empty.

@since WIP

Transformation

phull :: forall (a :: S -> Type) (s :: S). (POrd a, PIsData a) => Term s (PInterval a :--> (PInterval a :--> PInterval a))

phull i1 i2 gives the smallest interval that contains both i1 and i2.

Since: 2.1.1

pintersection :: forall (a :: S -> Type) (s :: S). (POrd a, PIsData a) => Term s (PInterval a :--> (PInterval a :--> PInterval a))

pintersection i1 i2 gives the largest interval that is contained in both i1 and i2.

Since: 2.1.1