Plutarch.LedgerApi.V1

Contexts

data PScriptPurpose (s :: S)

Since: 3.1.1

Constructors

PMinting (Term s (PDataRecord '["_0" ':= PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" ':= PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PCertifying (Term s (PDataRecord '["_0" ':= PDCert]))

newtype PScriptContext (s :: S)

Since: 3.1.1

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" ':= PTxInfo, "purpose" ':= PScriptPurpose]))

Certificates

data PDCert (s :: S)

Since: 3.1.1

Constructors

PDCertDelegRegKey (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PDCertDelegDeRegKey (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PDCertDelegDelegate (Term s (PDataRecord '["_0" ':= PStakingCredential, "_1" ':= PPubKeyHash]))

PDCertPoolRegister (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PPubKeyHash]))

PDCertPoolRetire (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PInteger]))

PDCertGenesis (Term s (PDataRecord '[]))

PDCertMir (Term s (PDataRecord '[]))

Credentials

data PCredential (s :: S)

Since: 2.0.0

Constructors

PPubKeyCredential (Term s (PDataRecord '["_0" ':= PPubKeyHash]))

PScriptCredential (Term s (PDataRecord '["_0" ':= PScriptHash]))

data PStakingCredential (s :: S)

Since: 2.0.0

Constructors

PStakingHash (Term s (PDataRecord '["_0" ':= PCredential]))

PStakingPtr (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PInteger, "_2" ':= PInteger]))

Value

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

newtype PLovelace (s :: S)

Since: 2.2.0

Constructors

PLovelace (Term s (PDataNewtype PInteger))

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s (PDataNewtype PByteString))

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s (PDataNewtype PByteString))

Time

newtype PPosixTime (s :: S)

Since: 2.0.0

Constructors

PPosixTime (Term s (PDataNewtype PInteger))

pposixTime :: forall (s :: S). Term s PInteger -> Term s PPosixTime

Construct a PPosixTime from a PInteger. Same as using the constructor, but a lot shorter.

@since WIP

unPPosixTime :: forall (s :: S). Term s PPosixTime -> Term s PInteger

Unwrap a PPosixTime to get a PInteger. Same as using pmatch, but a lot shorter. Also unwraps the Data encoding.

@since WIP

Intervals

data PExtended (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" ':= a]))

PPosInf (Term s (PDataRecord '[]))

newtype PLowerBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PLowerBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

newtype PUpperBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PUpperBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

newtype PInterval (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PInterval (Term s (PDataRecord '["from" ':= PLowerBound a, "to" ':= PUpperBound a]))

Script stuff

newtype PDatum (s :: S)

Since: 2.0.0

Constructors

PDatum (Term s PData)

newtype PRedeemer (s :: S)

Since: 2.0.0

Constructors

PRedeemer (Term s PData)

newtype PDatumHash (s :: S)

Since: 2.0.0

Constructors

PDatumHash (Term s (PDataNewtype PByteString))

newtype PRedeemerHash (s :: S)

Since: 2.0.0

Constructors

PRedeemerHash (Term s (PDataNewtype PByteString))

newtype PScriptHash (s :: S)

Since: 2.0.0

Constructors

PScriptHash (Term s (PDataNewtype PByteString))

Transactions

newtype PAddress (s :: S)

Since: 2.0.0

Constructors

PAddress (Term s (PDataRecord '["credential" ':= PCredential, "stakingCredential" ':= PMaybeData PStakingCredential]))

newtype PTxId (s :: S)

Hashed with BLAKE2b-256.

Since: 3.1.0

Constructors

PTxId (Term s (PDataRecord '["_0" ':= PByteString]))

newtype PTxOut (s :: S)

Since: 3.1.1

Constructors

PTxOut (Term s (PDataRecord '["address" ':= PAddress, "value" ':= PValue 'Sorted 'Positive, "datumHash" ':= PDatumHash]))

newtype PTxInInfo (s :: S)

Since: 3.1.1

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" ':= PTxOutRef, "resolved" ':= PTxOut]))

newtype PTxOutRef (s :: S)

Reference to a transaction output, with an index referencing which exact output we mean.

Since: 2.0.0

Constructors

PTxOutRef (Term s (PDataRecord '["id" ':= PTxId, "idx" ':= PInteger]))

newtype PPubKeyHash (s :: S)

Since: 2.0.0

Constructors

PPubKeyHash (Term s (PDataNewtype PByteString))

newtype PTxInfo (s :: S)

@since WIP

Constructors

PTxInfo (Term s (PDataRecord '["inputs" ':= PBuiltinList (PAsData PTxInInfo), "outputs" ':= PBuiltinList (PAsData PTxOut), "fee" ':= PValue 'Sorted 'Positive, "mint" ':= PValue 'Sorted 'NoGuarantees, "dCert" ':= PBuiltinList (PAsData PDCert), "wdrl" ':= PBuiltinList (PAsData (PBuiltinPair (PAsData PStakingCredential) (PAsData PInteger))), "validRange" ':= PInterval PPosixTime, "signatories" ':= PBuiltinList (PAsData PPubKeyHash), "data" ':= PBuiltinList (PAsData (PBuiltinPair (PAsData PDatumHash) (PAsData PDatum))), "id" ':= PTxId]))

Helpers

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Since: 2.0.0

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Since: 2.0.0

Constructors

Sorted

Unsorted

data Commutativity

Since: 2.0.0

Constructors

Commutative

NonCommutative

Utilities

Types

data PMaybeData (a :: S -> Type) (s :: S)

@since WIP

Constructors

PDJust (Term s (PAsData a))

PDNothing

newtype PRationalData s

A Rational type that corresponds to the data encoding used by Rational.

Since: 3.1.0

Constructors

PRationalData (Term s (PDataRecord '["numerator" ':= PInteger, "denominator" ':= PPositive]))

Utilities

pfromDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybeData a :--> a)

Extracts the element out of a PDJust and throws an error if its argument is PDNothing.

Since: 2.1.1

pisDJust :: forall (a :: PType) (s :: S). Term s (PMaybeData a :--> PBool)

Yield PTrue if a given PMaybeData is of the form PDJust _.

Since: 2.1.1

pmaybeData :: forall (a :: PType) (b :: PType) (s :: S). PIsData a => Term s (b :--> ((a :--> b) :--> (PMaybeData a :--> b)))

Special version of pmaybe that works with PMaybeData.

Since: 2.1.1

pdjust :: forall (a :: PType) (s :: S). PIsData a => Term s (a :--> PMaybeData a)

Construct a PDJust value.

Since: 2.1.1

pdnothing :: forall (a :: PType) (s :: S). Term s (PMaybeData a)

Construct a PDNothing value.

Since: 2.1.1

pmaybeToMaybeData :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybe a :--> PMaybeData a)

Construct a PMaybeData given a PMaybe. Could be useful if you want to "lift" from PMaybe to Maybe.

Since: 2.1.1

passertPDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PString :--> (PMaybeData a :--> a))

Extract the value stored in a PMaybeData container. If there's no value, throw an error with the given message.

Since: 2.1.1

prationalFromData :: ClosedTerm (PRationalData :--> PRational)

Since: 3.1.0

Orphan instances

PConstantDecl ScriptContext

Instance details

Associated Types

type PConstantRepr ScriptContext

type PConstanted ScriptContext :: PType

Methods

pconstantToRepr :: ScriptContext -> PConstantRepr ScriptContext

pconstantFromRepr :: PConstantRepr ScriptContext -> Maybe ScriptContext

PConstantDecl TxInInfo

Instance details

Associated Types

type PConstantRepr TxInInfo

type PConstanted TxInInfo :: PType

Methods

pconstantToRepr :: TxInInfo -> PConstantRepr TxInInfo

pconstantFromRepr :: PConstantRepr TxInInfo -> Maybe TxInInfo

PConstantDecl TxInfo

Instance details

Associated Types

type PConstantRepr TxInfo

type PConstanted TxInfo :: PType

Methods

pconstantToRepr :: TxInfo -> PConstantRepr TxInfo

pconstantFromRepr :: PConstantRepr TxInfo -> Maybe TxInfo

PConstantDecl TxOut

Instance details

Associated Types

type PConstantRepr TxOut

type PConstanted TxOut :: PType

Methods

pconstantToRepr :: TxOut -> PConstantRepr TxOut

pconstantFromRepr :: PConstantRepr TxOut -> Maybe TxOut