Plutarch.LedgerApi.V2

Contexts

data PScriptPurpose (s :: S)

Since: 3.1.1

Constructors

PMinting (Term s (PAsData PCurrencySymbol))

PSpending (Term s PTxOutRef)

PRewarding (Term s PStakingCredential)

PCertifying (Term s PDCert)

data PScriptContext (s :: S)

Since: 3.1.1

Constructors

PScriptContext

Fields

pscriptContext'txInfo :: Term s PTxInfo

pscriptContext'purpose :: Term s PScriptPurpose

Certificates

data PDCert (s :: S)

Since: 3.1.1

Constructors

PDCertDelegRegKey (Term s PStakingCredential)

PDCertDelegDeRegKey (Term s PStakingCredential)

PDCertDelegDelegate (Term s PStakingCredential) (Term s (PAsData PPubKeyHash))

PDCertPoolRegister (Term s (PAsData PPubKeyHash)) (Term s (PAsData PPubKeyHash))

PDCertPoolRetire (Term s (PAsData PPubKeyHash)) (Term s (PAsData PInteger))

PDCertGenesis

PDCertMir

Credentials

data PCredential (s :: S)

Since: 2.0.0

Constructors

PPubKeyCredential (Term s (PAsData PPubKeyHash))

PScriptCredential (Term s (PAsData PScriptHash))

data PStakingCredential (s :: S)

Since: 2.0.0

Constructors

PStakingHash (Term s PCredential)

PStakingPtr (Term s (PAsData PInteger)) (Term s (PAsData PInteger)) (Term s (PAsData PInteger))

Value

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

newtype PLovelace (s :: S)

Since: 2.2.0

Constructors

PLovelace (Term s PInteger)

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s PByteString)

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s PByteString)

Time

newtype PPosixTime (s :: S)

Since: 2.0.0

Constructors

PPosixTime (Term s PInteger)

pposixTime :: forall (s :: S). Term s PInteger -> Term s PPosixTime

Construct a PPosixTime from a PInteger. Same as using the constructor, but a lot shorter.

Since: 3.3.0

unPPosixTime :: forall (s :: S). Term s PPosixTime -> Term s PInteger

Unwrap a PPosixTime to get a PInteger. Same as using pmatch, but a lot shorter. Also unwraps the Data encoding.

Since: 3.3.0

Intervals

data PExtended (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PNegInf

PFinite (Term s (PAsData a))

PPosInf

data PLowerBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PLowerBound (Term s (PExtended a)) (Term s (PAsData PBool))

data PUpperBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PUpperBound (Term s (PExtended a)) (Term s (PAsData PBool))

data PInterval (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PInterval

Fields

pinteral'from :: Term s (PLowerBound a)

pinteral'to :: Term s (PUpperBound a)

Script stuff

newtype PDatum (s :: S)

Since: 2.0.0

Constructors

PDatum (Term s PData)

newtype PRedeemer (s :: S)

Since: 2.0.0

Constructors

PRedeemer (Term s PData)

newtype PDatumHash (s :: S)

Since: 2.0.0

Constructors

PDatumHash (Term s PByteString)

newtype PRedeemerHash (s :: S)

Since: 2.0.0

Constructors

PRedeemerHash (Term s PByteString)

newtype PScriptHash (s :: S)

Since: 2.0.0

Constructors

PScriptHash (Term s PByteString)

Transactions

data PAddress (s :: S)

Since: 2.0.0

Constructors

PAddress

Fields

paddress'credential :: Term s PCredential

paddress'stakingCredential :: Term s (PMaybeData PStakingCredential)

newtype PPubKeyHash (s :: S)

Since: 2.0.0

Constructors

PPubKeyHash (Term s PByteString)

newtype PTxId (s :: S)

Hashed with BLAKE2b-256.

Since: 3.1.0

Constructors

PTxId (Term s (PAsData PByteString))

data PTxInfo (s :: S)

Since: 3.1.1

Constructors

PTxInfo

Fields

ptxInfo'inputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))

ptxInfo'referenceInputs :: Term s (PAsData (PBuiltinList (PAsData PTxInInfo)))

ptxInfo'outputs :: Term s (PAsData (PBuiltinList (PAsData PTxOut)))

ptxInfo'fee :: Term s (PAsData (PValue 'Sorted 'Positive))

ptxInfo'mint :: Term s (PAsData (PValue 'Sorted 'NoGuarantees))

ptxInfo'dcert :: Term s (PAsData (PBuiltinList (PAsData PDCert)))

ptxInfo'wdrl :: Term s (PAsData (PMap 'Unsorted PStakingCredential PInteger))

ptxInfo'validRange :: Term s (PInterval PPosixTime)

ptxInfo'signatories :: Term s (PAsData (PBuiltinList (PAsData PPubKeyHash)))

ptxInfo'redeemers :: Term s (PAsData (PMap 'Unsorted PScriptPurpose PRedeemer))

ptxInfo'data :: Term s (PAsData (PMap 'Unsorted PDatumHash PDatum))

ptxInfo'id :: Term s PTxId

data PTxOut (s :: S)

Since: 2.0.0

Constructors

PTxOut

Fields

ptxOut'address :: Term s PAddress

ptxOut'value :: Term s (PAsData (PValue 'Sorted 'Positive))

ptxOut'datum :: Term s POutputDatum

ptxOut'referenceScript :: Term s (PMaybeData PScriptHash)

data PTxOutRef (s :: S)

Reference to a transaction output, with an index referencing which exact output we mean.

Since: 2.0.0

Constructors

PTxOutRef

Fields

ptxOutRef'id :: Term s PTxId

ptxOutRef'idx :: Term s (PAsData PInteger)

data PTxInInfo (s :: S)

Since: 3.1.1

Constructors

PTxInInfo

Fields

ptxInInfo'outRef :: Term s PTxOutRef

ptxInInfo'resolved :: Term s PTxOut

data POutputDatum (s :: S)

Since: 2.0.0

Constructors

PNoOutputDatum

POutputDatumHash

Fields

poutputDatum'datumHash :: Term s (PAsData PDatumHash)

POutputDatum
Inline datum as per CIP-0032

Fields

poutputDatum'outputDatum :: Term s PDatum

Helpers

newtype PMap (keysort :: KeyGuarantees) (k :: S -> Type) (v :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Since: 2.0.0

Constructors

Sorted

Unsorted

data Commutativity

Since: 2.0.0

Constructors

Commutative

NonCommutative

Utilities

Types

data PMaybeData (a :: S -> Type) (s :: S)

Since: 3.3.0

Constructors

PDJust (Term s (PAsData a))

PDNothing

data PRationalData s

A Rational type that corresponds to the data encoding used by Rational.

Since: 3.1.0

Constructors

PRationalData

Fields

prationalData'numerator :: Term s (PAsData PInteger)

prationalData'denominator :: Term s (PAsData PPositive)

Utilities

pfromDJust :: forall (a :: S -> Type) (s :: S). PIsData a => Term s (PMaybeData a :--> a)

Extracts the element out of a PDJust and throws an error if its argument is PDNothing.

Since: 2.1.1

pisDJust :: forall (a :: S -> Type) (s :: S). Term s (PMaybeData a :--> PBool)

Yield PTrue if a given PMaybeData is of the form PDJust _.

Since: 2.1.1

pmaybeData :: forall (a :: S -> Type) (b :: S -> Type) (s :: S). PIsData a => Term s (b :--> ((a :--> b) :--> (PMaybeData a :--> b)))

Special version of pmaybe that works with PMaybeData.

Since: 2.1.1

pdjust :: forall (a :: S -> Type) (s :: S). PIsData a => Term s (a :--> PMaybeData a)

Construct a PDJust value.

Since: 2.1.1

pdnothing :: forall (a :: S -> Type) (s :: S). Term s (PMaybeData a)

Construct a PDNothing value.

Since: 2.1.1

pmaybeToMaybeData :: forall (a :: S -> Type) (s :: S). PIsData a => Term s (PMaybe a :--> PMaybeData a)

Construct a PMaybeData given a PMaybe. Could be useful if you want to "lift" from PMaybe to Maybe.

Since: 2.1.1

passertPDJust :: forall (a :: S -> Type) (s :: S). PIsData a => Term s (PString :--> (PMaybeData a :--> a))

Extract the value stored in a PMaybeData container. If there's no value, throw an error with the given message.

Since: 2.1.1

prationalFromData :: ClosedTerm (PRationalData :--> PRational)

Since: 3.1.0