Plutarch.LedgerApi.V3

Description

Note

The PValue, PMap and PInterval-related functionality can be found in other modules, as these clash with the Plutarch prelude. These should be imported qualified.

Contexts

newtype PScriptContext (s :: S)

Since: 3.1.0

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" ':= PTxInfo, "redeemer" ':= PRedeemer, "scriptInfo" ':= PScriptInfo]))

newtype PTxInfo (s :: S)

Constructors

PTxInfo (Term s (PDataRecord '["inputs" ':= PBuiltinList (PAsData PTxInInfo), "referenceInputs" ':= PBuiltinList (PAsData PTxInInfo), "outputs" ':= PBuiltinList (PAsData PTxOut), "fee" ':= PLovelace, "mint" ':= PValue 'Sorted 'NonZero, "txCerts" ':= PBuiltinList (PAsData PTxCert), "wdrl" ':= PMap 'Unsorted PCredential PLovelace, "validRange" ':= PInterval PPosixTime, "signatories" ':= PBuiltinList (PAsData PPubKeyHash), "redeemers" ':= PMap 'Unsorted PScriptPurpose PRedeemer, "data" ':= PMap 'Unsorted PDatumHash PDatum, "id" ':= PTxId, "votes" ':= PMap 'Unsorted PVoter (PMap 'Unsorted PGovernanceActionId PVote), "proposalProcedures" ':= PBuiltinList (PAsData PProposalProcedure), "currentTreasuryAmount" ':= PMaybeData PLovelace, "treasuryDonation" ':= PMaybeData PLovelace]))

data PScriptInfo (s :: S)

Since: 3.1.0

Constructors

PMintingScript (Term s (PDataRecord '["_0" ':= PCurrencySymbol]))

PSpendingScript (Term s (PDataRecord '["_0" ':= PTxOutRef, "_1" ':= PMaybeData PDatum]))

PRewardingScript (Term s (PDataRecord '["_0" ':= PCredential]))

PCertifyingScript (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PTxCert]))

PVotingScript (Term s (PDataRecord '["_0" ':= PVoter]))

PProposingScript (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PProposalProcedure]))

data PScriptPurpose (s :: S)

Since: 2.0.0

Constructors

PMinting (Term s (PDataRecord '["_0" ':= PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" ':= PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" ':= PCredential]))
Since: 3.1.0

PCertifying (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PTxCert]))

PVoting (Term s (PDataRecord '["_0" ':= PVoter]))
Since: 3.1.0

PProposing (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PProposalProcedure]))
Since: 3.1.0

Tx

Types

newtype PTxOutRef (s :: S)

Since: 3.1.0

Constructors

PTxOutRef (Term s (PDataRecord '["id" ':= PTxId, "idx" ':= PInteger]))

newtype PTxOut (s :: S)

Since: 2.0.0

Constructors

PTxOut (Term s (PDataRecord '["address" ':= PAddress, "value" ':= PValue 'Sorted 'Positive, "datum" ':= POutputDatum, "referenceScript" ':= PMaybeData PScriptHash]))

newtype PTxId (s :: S)

Hashed with BLAKE2b-256.

Since: 3.1.0

Constructors

PTxId (Term s (PDataNewtype PByteString))

newtype PTxInInfo (s :: S)

An input of the pending transaction.

Since: 2.0.0

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" ':= PTxOutRef, "resolved" ':= PTxOut]))

data POutputDatum (s :: S)

Since: 2.0.0

Constructors

PNoOutputDatum (Term s (PDataRecord '[]))

POutputDatumHash (Term s (PDataRecord '["datumHash" ':= PDatumHash]))

POutputDatum (Term s (PDataRecord '["outputDatum" ':= PDatum]))
Inline datum as per CIP-0032

Functions

pgetContinuingOutputs :: forall (s :: S). Term s (PBuiltinList PTxInInfo :--> (PBuiltinList PTxOut :--> (PTxOutRef :--> PBuiltinList PTxOut)))

Find the output txns corresponding to the input being validated.

Takes as arguments the inputs, outputs and the spending transaction referenced from PScriptPurpose.

Example:

 ctx <- tcont $ pletFields ["txInfo", "purpose"] sc
 pmatchC (getField "purpose" ctx) >>= case
   PSpending outRef' -> do
     let outRef = pfield "_0" # outRef'
         inputs = pfield "inputs" # (getField "txInfo" ctx)
         outputs = pfield "outputs" # (getField @"txInfo" ctx)
     pure $ pgetContinuingOutputs # inputs # outputs # outRef
   _ ->
     pure $ ptraceInfoError "not a spending tx"
 

Since: 2.1.0

pfindOwnInput :: forall (s :: S). Term s (PBuiltinList PTxInInfo :--> (PTxOutRef :--> PMaybe PTxInInfo))

Find the input being spent in the current transaction.

Takes as arguments the inputs, as well as the spending transaction referenced from PScriptPurpose.

Example:

 ctx <- tcont $ pletFields ["txInfo", "purpose"] sc
 pmatchC (getField "purpose" ctx) >>= case
   PSpending outRef' -> do
     let outRef = pfield "_0" # outRef'
         inputs = pfield "inputs" # (getField @"txInfo" ctx)
     pure $ pfindOwnInput # inputs # outRef
   _ ->
     pure $ ptraceInfoError "not a spending tx"
 

Since: 2.1.0

Script

Types

newtype PDatum (s :: S)

Since: 2.0.0

Constructors

PDatum (Term s PData)

newtype PDatumHash (s :: S)

Since: 2.0.0

Constructors

PDatumHash (Term s (PDataNewtype PByteString))

newtype PRedeemer (s :: S)

Since: 2.0.0

Constructors

PRedeemer (Term s PData)

newtype PRedeemerHash (s :: S)

Since: 2.0.0

Constructors

PRedeemerHash (Term s (PDataNewtype PByteString))

newtype PScriptHash (s :: S)

Since: 2.0.0

Constructors

PScriptHash (Term s (PDataNewtype PByteString))

Functions

scriptHash :: Script -> ScriptHash

Hash a script, appending the Plutus V2 prefix.

Since: 2.0.0

datumHash :: Datum -> DatumHash

Since: 2.0.0

redeemerHash :: Redeemer -> RedeemerHash

Since: 2.0.0

dataHash :: forall (a :: Type). ToData a => a -> BuiltinByteString

Since: 2.0.0

pparseDatum :: forall (a :: S -> Type) (s :: S). PTryFrom PData (PAsData a) => Term s (PDatumHash :--> (PMap 'Unsorted PDatumHash PDatum :--> PMaybe (PAsData a)))

Lookup up the datum given the datum hash.

Takes as argument the datum assoc list from a PTxInfo. Validates the datum using PTryFrom.

Example:

 pparseDatum MyType # datumHash #$ pfield "datums" # txinfo
 

Since: 2.1.2

Value

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s (PDataNewtype PByteString))

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s (PDataNewtype PByteString))

newtype PLovelace (s :: S)

Since: 2.2.0

Constructors

PLovelace (Term s (PDataNewtype PInteger))

Assoc map

Types

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Since: 2.0.0

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Since: 2.0.0

Constructors

Sorted

Unsorted

data Commutativity

Since: 2.0.0

Constructors

Commutative

NonCommutative

Address

data PCredential (s :: S)

Since: 2.0.0

Constructors

PPubKeyCredential (Term s (PDataRecord '["_0" ':= PPubKeyHash]))

PScriptCredential (Term s (PDataRecord '["_0" ':= PScriptHash]))

data PStakingCredential (s :: S)

Since: 2.0.0

Constructors

PStakingHash (Term s (PDataRecord '["_0" ':= PCredential]))

PStakingPtr (Term s (PDataRecord '["_0" ':= PInteger, "_1" ':= PInteger, "_2" ':= PInteger]))

newtype PAddress (s :: S)

Since: 2.0.0

Constructors

PAddress (Term s (PDataRecord '["credential" ':= PCredential, "stakingCredential" ':= PMaybeData PStakingCredential]))

Time

newtype PPosixTime (s :: S)

Since: 2.0.0

Constructors

PPosixTime (Term s (PDataNewtype PInteger))

pposixTime :: forall (s :: S). Term s PInteger -> Term s PPosixTime

Construct a PPosixTime from a PInteger. Same as using the constructor, but a lot shorter.

@since WIP

unPPosixTime :: forall (s :: S). Term s PPosixTime -> Term s PInteger

Unwrap a PPosixTime to get a PInteger. Same as using pmatch, but a lot shorter. Also unwraps the Data encoding.

@since WIP

Interval

newtype PInterval (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PInterval (Term s (PDataRecord '["from" ':= PLowerBound a, "to" ':= PUpperBound a]))

newtype PLowerBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PLowerBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

newtype PUpperBound (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PUpperBound (Term s (PDataRecord '["_0" ':= PExtended a, "_1" ':= PBool]))

data PExtended (a :: S -> Type) (s :: S)

Since: 2.0.0

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" ':= a]))

PPosInf (Term s (PDataRecord '[]))

CIP-1694

data PTxCert (s :: S)

Since: 3.1.0

Constructors

PTxCertRegStaking (Term s (PDataRecord '["_0" ':= PCredential, "_1" ':= PMaybeData PLovelace]))

PTxCertUnRegStaking (Term s (PDataRecord '["_0" ':= PCredential, "_1" ':= PMaybeData PLovelace]))

PTxCertDelegStaking (Term s (PDataRecord '["_0" ':= PCredential, "_1" ':= PDelegatee]))

PTxCertRegDeleg (Term s (PDataRecord '["_0" ':= PCredential, "_1" ':= PDelegatee, "_2" ':= PLovelace]))

PTxCertRegDRep (Term s (PDataRecord '["_0" ':= PDRepCredential, "_1" ':= PLovelace]))

PTxCertUpdateDRep (Term s (PDataRecord '["_0" ':= PDRepCredential]))

PTxCertUnRegDRep (Term s (PDataRecord '["_0" ':= PDRepCredential, "_1" ':= PLovelace]))

PTxCertPoolRegister (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PPubKeyHash]))

PTxCertPoolRetire (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PInteger]))

PTxCertAuthHotCommittee (Term s (PDataRecord '["_0" ':= PColdCommitteeCredential, "_1" ':= PHotCommitteeCredential]))

PTxCertResignColdCommittee (Term s (PDataRecord '["_0" ':= PColdCommitteeCredential]))

data PDelegatee (s :: S)

Since: 3.1.0

Constructors

PDelegStake (Term s (PDataRecord '["_0" ':= PPubKeyHash]))

PDelegVote (Term s (PDataRecord '["_0" ':= PDRep]))

PDelegStakeVote (Term s (PDataRecord '["_0" ':= PPubKeyHash, "_1" ':= PDRep]))

newtype PDRepCredential (s :: S)

Since: 3.1.0

Constructors

PDRepCredential (Term s PCredential)

newtype PColdCommitteeCredential (s :: S)

Since: 3.1.0

Constructors

PColdCommitteeCredential (Term s PCredential)

newtype PHotCommitteeCredential (s :: S)

Since: 3.1.0

Constructors

PHotCommitteeCredential (Term s PCredential)

data PDRep (s :: S)

Since: 3.1.0

Constructors

PDRep (Term s (PDataRecord '["_0" ':= PDRepCredential]))

PDRepAlwaysAbstain (Term s (PDataRecord '[]))

PDRepAlwaysNoConfidence (Term s (PDataRecord '[]))

data PVoter (s :: S)

Since: 3.1.0

Constructors

PCommitteeVoter (Term s (PDataRecord '["_0" ':= PHotCommitteeCredential]))

PDRepVoter (Term s (PDataRecord '["_0" ':= PDRepCredential]))

PStakePoolVoter (Term s (PDataRecord '["_0" ':= PPubKeyHash]))

newtype PGovernanceActionId (s :: S)

Since: 3.1.0

Constructors

PGovernanceActionId (Term s (PDataRecord '["txId" ':= PTxId, "govActionIx" ':= PInteger]))

data PVote (s :: S)

Since: 3.1.0

Constructors

PVoteYes (Term s (PDataRecord '[]))

PVoteNo (Term s (PDataRecord '[]))

PAbstain (Term s (PDataRecord '[]))

newtype PProtocolVersion (s :: S)

Since: 3.1.0

Constructors

PProtocolVersion (Term s (PDataRecord '["major" ':= PInteger, "minor" ':= PInteger]))

newtype PProposalProcedure (s :: S)

Since: 3.1.0

Constructors

PProposalProcedure (Term s (PDataRecord '["deposit" ':= PLovelace, "returnAddr" ':= PCredential, "governanceAction" ':= PGovernanceAction]))

data PGovernanceAction (s :: S)

Since: 3.1.0

Constructors

PParameterChange (Term s (PDataRecord '["_0" ':= PMaybeData PGovernanceActionId, "_1" ':= PChangedParameters, "_2" ':= PMaybeData PScriptHash]))

PHardForkInitiation (Term s (PDataRecord '["_0" ':= PMaybeData PGovernanceActionId, "_1" ':= PProtocolVersion]))

PTreasuryWithdrawals (Term s (PDataRecord '["_0" ':= PMap 'Unsorted PCredential PLovelace, "_1" ':= PMaybeData PScriptHash]))

PNoConfidence (Term s (PDataRecord '["_0" ':= PMaybeData PGovernanceActionId]))

PUpdateCommittee (Term s (PDataRecord '["_0" ':= PMaybeData PGovernanceActionId, "_1" ':= PBuiltinList (PAsData PColdCommitteeCredential), "_2" ':= PMap 'Unsorted PColdCommitteeCredential PInteger, "_3" ':= PRationalData]))

PNewConstitution (Term s (PDataRecord '["_0" ':= PMaybeData PGovernanceActionId, "_1" ':= PConstitution]))

PInfoAction (Term s (PDataRecord '[]))

newtype PChangedParameters (s :: S)

Since: 3.1.0

Constructors

PChangedParameters (Term s PData)

newtype PConstitution (s :: S)

A constitution, omitting the optional anchor.

Since: 3.1.0

Constructors

PConstitution (Term s (PDataRecord '["_0" := PMaybeData PScriptHash]))

newtype PCommittee (s :: S)

Since: 3.1.0

Constructors

PCommittee (Term s (PDataRecord '["members" ':= PMap 'Unsorted PColdCommitteeCredential PInteger, "quorum" ':= PRationalData]))

Crypto

Types

newtype PubKey

Since: 2.0.0

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

Since: 2.0.0

newtype PPubKeyHash (s :: S)

Since: 2.0.0

Constructors

PPubKeyHash (Term s (PDataNewtype PByteString))

pubKeyHash :: PubKey -> PubKeyHash

Since: 2.0.0

Utilities

Types

data PMaybeData (a :: S -> Type) (s :: S)

@since WIP

Constructors

PDJust (Term s (PAsData a))

PDNothing

newtype PRationalData s

A Rational type that corresponds to the data encoding used by Rational.

Since: 3.1.0

Constructors

PRationalData (Term s (PDataRecord '["numerator" ':= PInteger, "denominator" ':= PPositive]))

Utilities

pfromDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybeData a :--> a)

Extracts the element out of a PDJust and throws an error if its argument is PDNothing.

Since: 2.1.1

pisDJust :: forall (a :: PType) (s :: S). Term s (PMaybeData a :--> PBool)

Yield PTrue if a given PMaybeData is of the form PDJust _.

Since: 2.1.1

pmaybeData :: forall (a :: PType) (b :: PType) (s :: S). PIsData a => Term s (b :--> ((a :--> b) :--> (PMaybeData a :--> b)))

Special version of pmaybe that works with PMaybeData.

Since: 2.1.1

pdjust :: forall (a :: PType) (s :: S). PIsData a => Term s (a :--> PMaybeData a)

Construct a PDJust value.

Since: 2.1.1

pdnothing :: forall (a :: PType) (s :: S). Term s (PMaybeData a)

Construct a PDNothing value.

Since: 2.1.1

pmaybeToMaybeData :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybe a :--> PMaybeData a)

Construct a PMaybeData given a PMaybe. Could be useful if you want to "lift" from PMaybe to Maybe.

Since: 2.1.1

passertPDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PString :--> (PMaybeData a :--> a))

Extract the value stored in a PMaybeData container. If there's no value, throw an error with the given message.

Since: 2.1.1

prationalFromData :: ClosedTerm (PRationalData :--> PRational)

Since: 3.1.0