Plutarch.LedgerApi.Value

Description

Value-related functionality. In order to keep the interface efficient and safe at the same time, there is a type-level distinction between PValues that are guaranteed to be properly normalized and those that provide no such guarantee.

Also for efficiency reasons, the Ada-specific functions assume that there can be only one token name for the Ada currency symbol, and they don't check whether it matches adaToken.

Types

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Since: 2.0.0

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

newtype PCurrencySymbol (s :: S)

Since: 2.0.0

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Since: 2.0.0

Constructors

PTokenName (Term s PByteString)

data AmountGuarantees

Since: 2.0.0

Constructors

NoGuarantees

NonZero

Positive

newtype PLovelace (s :: S)

Since: 2.2.0

Constructors

PLovelace (Term s PInteger)

newtype PAssetClass (s :: S)

Since: 3.3.0

Constructors

PAssetClass (Term s (PBuiltinPair (PAsData PCurrencySymbol) (PAsData PTokenName)))

Functions

Creation

PCurrencySymbol

padaSymbol :: forall (s :: S). Term s PCurrencySymbol

The PCurrencySymbol of the Ada currency.

Since: 2.1.1

padaSymbolData :: forall (s :: S). Term s (PAsData PCurrencySymbol)

Data-encoded PCurrencySymbol of the Ada currency.

Since: 2.1.1

PTokenName

padaToken :: Term s PTokenName

The PTokenName of the Ada currency.

Since: 2.1.1

PValue

psingleton :: forall (s :: S). Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PValue 'Sorted 'NonZero)))

Construct a singleton PValue containing only the given quantity of the given currency.

Since: 2.1.1

psingletonData :: forall (s :: S). Term s (PAsData PCurrencySymbol :--> (PAsData PTokenName :--> (PAsData PInteger :--> PValue 'Sorted 'NonZero)))

Construct a singleton PValue containing only the given quantity of the given currency, taking data-encoded parameters.

Since: 2.1.1

pconstantPositiveSingleton :: forall (s :: S). (forall (s' :: S). Term s' PCurrencySymbol) -> (forall (s' :: S). Term s' PTokenName) -> (forall (s' :: S). Term s' PInteger) -> Term s (PValue 'Sorted 'Positive)

Construct a constant singleton PValue containing only the given positive quantity of the given currency.

Since: 2.1.1

Transformation

passertPositive :: forall (kg :: KeyGuarantees) (ag :: AmountGuarantees) (s :: S). Term s (PValue kg ag :--> PValue kg 'Positive)

Given a PValue, either construct another PValue with the same contents and a proof that all amounts in it are positive, or error.

Since: 2.0.0

passertNonZero :: forall (kg :: KeyGuarantees) (ag :: AmountGuarantees). forall (s :: S). Term s (PValue kg ag :--> PValue kg 'NonZero)

passertSorted :: forall (anyKey :: KeyGuarantees) (anyAmount :: AmountGuarantees) (s :: S). Term s (PValue anyKey anyAmount :--> PValue 'Sorted 'NonZero)

Assert the value is properly sorted and normalized.

Since: 2.1.1

pforgetPositive :: forall (a :: AmountGuarantees) (k :: KeyGuarantees) (s :: S). Term s (PValue k 'Positive) -> Term s (PValue k a)

'Forget' that a Value has an only-positive guarantee.

Since: 2.0.0

pforgetSorted :: forall (a :: AmountGuarantees) (k :: KeyGuarantees) (s :: S). Term s (PValue 'Sorted a) -> Term s (PValue k a)

Forget the knowledge of all value's guarantees.

Since: 2.1.1

pnormalize :: forall (any :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted any :--> PValue 'Sorted 'NonZero)

Normalize the argument to contain no zero quantity nor empty token map.

Since: 2.0.0

padaOnlyValue :: forall (v :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted v :--> PValue 'Sorted v)

Strip all non-Ada from a PValue.

Since: 2.1.1

pnoAdaValue :: forall (v :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted v :--> PValue 'Sorted v)

Strip all Ada from a PValue.

Since: 2.1.1

Partial ordering

pltPositive :: forall (s :: S). Term s (PValue 'Sorted 'Positive) -> Term s (PValue 'Sorted 'Positive) -> Term s PBool

Mimics the lt operation on plutus-ledger-api's Value.

Since: 3.3.0

pltNonZero :: forall (s :: S). Term s (PValue 'Sorted 'NonZero) -> Term s (PValue 'Sorted 'NonZero) -> Term s PBool

As pltPositive, but for nonzero guaranteed PValues instead.

Since: 3.3.0

pleqPositive :: forall (s :: S). Term s (PValue 'Sorted 'Positive) -> Term s (PValue 'Sorted 'Positive) -> Term s PBool

Mimics the leq operation on plutus-ledger-api's Value.

Since: 3.3.0

pleqNonZero :: forall (s :: S). Term s (PValue 'Sorted 'NonZero) -> Term s (PValue 'Sorted 'NonZero) -> Term s PBool

As pletPositive, but for nonzero guaranteed PValues instead.

Since: 3.3.0

pcheckBinRel :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Term s ((PInteger :--> (PInteger :--> PBool)) :--> (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PBool)))

Given a description of a relation on amounts, check whether that relation holds over sorted PValues.

Important note

This is intended for use with boolean comparison functions, which must define at least a partial order (total orders and equivalences are acceptable as well). Use of this with anything else is not guaranteed to give anything resembling a sensible answer. Use with extreme care.

Since: 2.0.0

Combination

pleftBiasedCurrencyUnion :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PValue 'Sorted 'NoGuarantees))

Combine two PValues, taking the tokens from the left only, if a currency occurs on both sides.

Since: 2.1.1

pleftBiasedTokenUnion :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PValue 'Sorted 'NoGuarantees))

Combine two PValues, taking the tokens from the left only, if a token name of the same currency occurs on both sides.

Prefer this over 'punionResolvingCollisionsWith NonCommutative # plam const'. It is equivalent, but performs better.

Since: 2.1.1

punionResolvingCollisionsWith :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Commutativity -> Term s ((PInteger :--> (PInteger :--> PInteger)) :--> (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PValue 'Sorted 'NoGuarantees)))

Combine two PValues applying the given function to any pair of quantities with the same asset class. Note that the result is _not_ normalized and may contain zero quantities.

Since: 2.0.0

punionResolvingCollisionsWithData :: forall (any0 :: AmountGuarantees) (any1 :: AmountGuarantees) (s :: S). Commutativity -> Term s ((PAsData PInteger :--> (PAsData PInteger :--> PAsData PInteger)) :--> (PValue 'Sorted any0 :--> (PValue 'Sorted any1 :--> PValue 'Sorted 'NoGuarantees)))

Combine two PValues applying the given function to any pair of data-encoded quantities with the same asset class. Note that the result is _not_ normalized and may contain zero quantities.

Since: 2.1.1

Queries

pvalueOf :: forall (anyKey :: KeyGuarantees) (anyAmount :: AmountGuarantees) (s :: S). Term s (PValue anyKey anyAmount :--> (PCurrencySymbol :--> (PTokenName :--> PInteger)))

Get the quantity of the given currency in the PValue.

Since: 2.1.1

plovelaceValueOf :: forall (v :: AmountGuarantees) (s :: S). Term s (PValue 'Sorted v :--> PInteger)

Get the amount of Lovelace in the PValue.

Since: 2.1.1

pisAdaOnlyValue :: forall (s :: S). Term s (PValue 'Sorted 'Positive :--> PBool)

Test if the value contains nothing but Ada

Since: 2.1.1