Plutarch.LedgerApi.Utils

Description

Useful tools that aren't part of the Plutarch API per se, but get used in multiple places.

Types

newtype Mret (a :: S -> Type) (s :: S)

Term, but with its type arguments flipped. This is a useful helper for defining PTryFrom instances.

For example, consider the PTryFrom instance for PTokenName:

instance PTryFrom PData (PAsData PTokenName) where
   type PTryFromExcess PData (PAsData PTokenName) = Mret PTokenName

We need to do this because PTryFromExcess expects something of kind S -> Type, but Term has kind S -> (S -> Type) -> Type, which doesn't quite fit. By using Mret, we end up with something of kind (S -> Type) -> S -> Type, which fits.

The name is just Term written backwards.

Since: 2.0.0

Constructors

Mret (Term s a)

data PMaybeData (a :: S -> Type) (s :: S)

@since WIP

Constructors

PDJust (Term s (PAsData a))

PDNothing

newtype PRationalData s

A Rational type that corresponds to the data encoding used by Rational.

Since: 3.1.0

Constructors

PRationalData (Term s (PDataRecord '["numerator" ':= PInteger, "denominator" ':= PPositive]))

Functions

PMaybeData

pfromDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybeData a :--> a)

Extracts the element out of a PDJust and throws an error if its argument is PDNothing.

Since: 2.1.1

pisDJust :: forall (a :: PType) (s :: S). Term s (PMaybeData a :--> PBool)

Yield PTrue if a given PMaybeData is of the form PDJust _.

Since: 2.1.1

pmaybeData :: forall (a :: PType) (b :: PType) (s :: S). PIsData a => Term s (b :--> ((a :--> b) :--> (PMaybeData a :--> b)))

Special version of pmaybe that works with PMaybeData.

Since: 2.1.1

pdjust :: forall (a :: PType) (s :: S). PIsData a => Term s (a :--> PMaybeData a)

Construct a PDJust value.

Since: 2.1.1

pdnothing :: forall (a :: PType) (s :: S). Term s (PMaybeData a)

Construct a PDNothing value.

Since: 2.1.1

pmaybeToMaybeData :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybe a :--> PMaybeData a)

Construct a PMaybeData given a PMaybe. Could be useful if you want to "lift" from PMaybe to Maybe.

Since: 2.1.1

pmaybeDataToMaybe :: forall (a :: PType) (s :: S). PIsData a => Term s (PMaybeData a :--> PMaybe a)

Inverse of pmaybeToMaybeData

@since WIP

passertPDJust :: forall (a :: PType) (s :: S). PIsData a => Term s (PString :--> (PMaybeData a :--> a))

Extract the value stored in a PMaybeData container. If there's no value, throw an error with the given message.

Since: 2.1.1

PRationalData

prationalFromData :: ClosedTerm (PRationalData :--> PRational)

Since: 3.1.0

Orphan instances

PConstantDecl Rational

Instance details

Associated Types

type PConstantRepr Rational

type PConstanted Rational :: PType

Methods

pconstantToRepr :: Rational -> PConstantRepr Rational

pconstantFromRepr :: PConstantRepr Rational -> Maybe Rational

PConstantData a => PConstantDecl (Maybe a)

Instance details

Associated Types

type PConstantRepr (Maybe a)

type PConstanted (Maybe a) :: PType

Methods

pconstantToRepr :: Maybe a -> PConstantRepr (Maybe a)

pconstantFromRepr :: PConstantRepr (Maybe a) -> Maybe (Maybe a)